### 2023-01-06 #5043 @editorialbot
[PRE REVIEW]: PII-Codex: a Python library for PII detection, categorization, and severity assessment
**Submitting author:** @EdyVision (<a href="http://orcid.org/0000-0003-0665-098X">Eidan Rosado</a>)
**Repository:** https://github.com/EdyVision/pii-codex
**Branch with paper.md** (empty if default branch): joss-paper-submission
**Version:** 0.4.3
