### 2023-01-07 1aee6b6
Added JOSS badge to readme
