### 2023-01-29 #5043 @arfon
@EdyVision – I see there are two licenses in this repository, BSD and HL3. The former is OSI-approved, the later is not. As an OSI-affiliated project, JOSS is committed to the OSI definition of _open source_. I believe this is the first time we've had this situation arise and will have to consult with the OSI team about this. As such, I'm going to mark this submission as paused for now.
