### 2023-02-01 d669e7d
Removed HL3 license per JOSS requirement and added description in pyproject.toml
