+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "pii-codex"
 version = "0.4.3"
-description = ""
+description = "PII Detection, Categorization, and Severity Assessment"
 authors = ["Eidan J. Rosado"]
 license = "BSD 3-Clause"
 readme = "README.md"
