### 2023-02-01 #5043 @EdyVision
> @EdyVision – having consulted on this, I'm afraid we can't review this submission in its current form as the HL3 license isn't OSI-approved, and conflicts with the BSD one. As an OSI-affiliate project, JOSS is committed to maintaining the [OSD definition](https://opensource.org/osd) of open source.
>
> I think there are two options here:
>
> 1. Remove the HL3 license.
> 2. Withdraw the submission.
>
> Which would you like to do?

I went ahead and removed the HL3 license.  Thank you for that and I apologize for the mix-up.
