### 2023-02-01 #5043 @arfon
> I went ahead and removed the HL3 license @arfon. Thank you for that and I apologize for the mix-up.

Many thanks. Could you also remove from the `main` branch?
