### 2023-02-01 4d64461
Removal of H3L License per JOSS Requirements (#29)
