### 2023-02-21 #5043 @EdyVision
> @EdyVision if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Thank you @arfon , the list is filled with so many people! That's exciting. At a glance at the background and their desired languages and topics of interest, I was thinking of `gradvohl`, `a2un`, `ma-sadeghi`, `VenkteshV`, `natematias`, or `ruchirtewari`. This group is a mix of cybersecurity, NLP, computational social science, and scientific computing. Happy to keep looking if they are unavailable of course. Thanks again!
