### 2023-02-25 #5043 @arfon
:wave: @ruchirtewari @natematias @VenkteshV would any of you be willing to review this submission for JOSS? The submission under consideration is PII-Codex: a Python library for PII detection, categorization, and severity assessment (https://github.com/EdyVision/pii-codex)

Many thanks
Arfon
