### 2023-03-29 #5043 @gradvohl
Hello @arfon,

As I mentioned in a previous message, I will be available in early April, probably around April 10th (after Easter). Perhaps, JOSS has other reviewers who can examine the article before then. At least, that's what it seemed to me when they declined my application for the editors' group, declaring that there was already enough editorial capacity in the area I applied for. If @EdyVision can wait until then, I will be glad to review the submission after April 10th.

Best regards.
