### 2023-04-23 #5402 @gradvohl
# Reviewer report for JOSS paper #5042
**Paper title**: PII-Codex: a Python library for PII detection, categorization, and severity assessment

## Review criteria
**Note**: Whenever I marked an item with :warning: or :no_entry:, you can get more information about that evaluation in the [**Reviewer Comments**](#reviewer-comments) section.

* A list of the authors of the software: :no_entry: -- Please see the [**About authors and their affiliation**](#about-authors-and-their-affiliation)  in the [**Reviewer Comments**](#reviewer-comments) section.
* Author affiliations: :no_entry:-- Please see the [**About authors and their affiliation**](#about-authors-and-their-affiliation)  in the [**Reviewer Comments**](#reviewer-comments) section.
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :warning:-- Please see the [**About the statement of need**](#about-the-statement-of-need)  in the [**Reviewer Comments**](#reviewer-comments) section.
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :warning:-- Please see the [**Mentions of any ongoing research projects**](#mentions-of-any-ongoing-research-projects) in the [**Reviewer Comments**](#reviewer-comments) section).
* A list of key references, including a link to the software archive:  :no_entry: -- Please see the [**A list of key references**](#a-list-of-key-references)   in the [**Reviewer Comments**](#reviewer-comments) section.

## Review items
* Software license: :thumbsup: (good -- BSD-3-Clause)
* Substantial scholarly effort: :wave: :warning: (**sort of** -- Please see the [**Scholarly effort**](#scholarly-effort)  in the [**Reviewer Comments**](#reviewer-comments) section)
* Documentation:
    * A statement of need: :wave: (**sort of**) :warning:-- Please see the [**About the statement of need**](#about-the-statement-of-need)   in the [**Reviewer Comments**](#reviewer-comments) section).
    * Installation instructions: (**bad** -- Please see the [**About the Installation Instructions**](#about-the-installation-instructions)   in the [**Reviewer Comments**](#reviewer-comments) section).
    * Example usage: :wave: (**sort of** -- Please see the [**Example usage**](#example-usage) in the [**Reviewer Comments**](#reviewer-comments) section).
    * API documentation: :wave: (**sort of** -- Please see the [**API Docs**](#api-docs)  in the [**Reviewer Comments**](#reviewer-comments) section).
    * Community guidelines:  :thumbsdown: (**bad** -- Please see the [**Community guidelines**](#community-guidelines)   in the [**Reviewer Comments**](#reviewer-comments) section).
    * Functionality: :thumbsdown: (**bad** -- Please see the [**Tests on source code**](#tests-on-source-code)  in the [**Reviewer Comments**](#reviewer-comments) section).
    * Tests: :thumbsdown: (**bad** -- Please see the [**Tests on source code**](#tests-on-source-code)   in the [**Reviewer Comments**](#reviewer-comments) section).
    *
* Software paper
	* The file ``paper.md`` file is in the ``joss-paper-submission`` branch. [:no_entry:](#about-paper.md-file).
    * Authors: :thumbsup: (**good**)
    * A statement of need: :thumbsup: (**good**)
    * Quality of writing: :thumbsup: (**good**)
    * References: :wave: (**sort of**) -- Please see the [**A list of key references**](#a-list-of-key-references)  section in the [**Reviewer Comments**](#reviewer-comments) section.

## Test environment
We try to run the software on Rocky Linux X86_64 computer, but the installation fails.

## Reviewer comments
The following comments refer to items the author could improve to meet JOSS submission requirements.

### About authors and their affiliation
The repository needs to clearly show who the authors are and their affiliation in the ``README.md`` file.

### About the statement of need
The  ``README.md``
needs to state clearly what problems the author designate the software to solve and who the target audience is. Although the first section of the ``README.md`` file quickly mentions the software's importance, the author must make it explicit and meaningful. Therefore, I suggest a special section in the ``README.md`` file with the statement of need.

### About the Installation Instructions
When I followed the installation instructions using the command ``pip3 install -i pii-codex``, it raised the following error:
```
You must give at least one requirement to install (see "pip help install")
```
After that, when I remove the flag  ``-i``, the following error message pops up:
```
  Could not find a version that satisfies the requirement pii-codex (from versions: )
No matching distribution found for pii-codex
```
The installation instructions must work as smoothly as possible.

### About ``paper.md`` file
I did not find the ``paper.md`` file in the main branch. I consider it better to put ``paper.md`` and ``paper.bib`` in the master branch.

### Mentions of any ongoing research projects
There are mentions of other research that influenced the software. However, it would be interesting if the authors informed the readers if any project that uses the PII-Codex exists. I recommend citing that project or, perhaps, writing about its potential in other research projects.

### A list of key references
There is no list of key references in the repository, just in the ``paper.md`` file. Although the ``README.md`` file cites two references -- Milne et al. (2016) and Schwartz and Solove (2012) --, there is no information about those references in the repository.
In addition, most of the references the author cited in the paper point to manuals or reports. That situation is acceptable, but I miss more citations of scientific literature.

### Example usage
The example usage described in the ``README.md`` file is very straightforward. However, there is a ``docs`` directory with more specific information, which is good. Nevertheless, the first section (_PII Type Mappings_) has a broken link.

### Scholarly effort
Reading the documentation and the ``README.md`` file, the substantial effort to build the software needs to be clarified. I highlight that my opinion is not a demerit of the software. Perhaps the author can present the scholarly effort to build the software more clearly, citing more studies and references.

### API Docs
I did not find any documentation about the code regarding the API documentation. The source code has some documentation strings, but it needs better documentation. Therefore, I strongly recommend using a standard to document the source code. I suggest using the [Sphinx](https://www.sphinx-doc.org/en/master/tutorial/describing-code.html) software. Sphinx can generate the docs (in HTML or PDF) from the documented source files.

### Community guidelines
There is not enough information about contributing to PII-Index and there are no community guidelines. Therefore, I propose creating a **Contributing** section (or a ``CONTRIBUTING.md`` file) with more specific instructions about contributing, including contact channels with main developers.

### Tests on source code
We cannot test the software as we cannot install it using the proposed instructions.

## Conclusion
Analyzing all the aspects mentioned earlier, I rank the paper as **major revisions** needed. Although most of those revisions are very easy to implement, I consider the lack of documentation, the absence of the ``paper.md`` file in the master branch, and the failure of simple installation instructions in the ``README.md`` prevents acceptance of the paper at this time.
