+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pii-codex"
-version = "0.4.3"
+version = "0.4.4"
 description = ""
 authors = ["Eidan J. Rosado"]
 license = "BSD 3-Clause"
@@ -13,7 +13,7 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: BSD License",
     "Programming Language :: Python :: 3", ]
-include = ["LICENSE", "LICENSE-ETHICS"]
+include = ["LICENSE"]
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.11"
@@ -24,9 +24,9 @@ pillow = "^9.3.0"  # security dependency
 
 # Optional items for analyses - users have the option to use the integrated Presidio analyzer or
 # avoid installing the following and bring their own detector, using the adapters
-spacy = { version = "^3.4.1", optional = true }
-presidio-analyzer = { version = "^2.2.31", optional = true }
-presidio-anonymizer = { version ="^2.2.31", optional = true }
+spacy = { version = "^3.5.2", optional = true }
+presidio-analyzer = { version = "^2.2.32", optional = true }
+presidio-anonymizer = { version ="^2.2.32", optional = true }
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.2.0"
@@ -43,7 +43,7 @@ jupyter_core = "^4.11.2"
 importlib-resources = "^5.9.0"
 seaborn = "^0.12.0"
 pre-commit = "^2.20.0"
-en_core_web_lg = {url="https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0.tar.gz#egg=en_core_web_lg"}
+en_core_web_lg = {url="https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0.tar.gz#egg=en_core_web_lg"}
 
 [tool.poetry.extras]
 detections = [
@@ -83,5 +83,5 @@ log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(line
 log_file_date_format = "%Y-%m-%d %H:%M:%S"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["poetry>=1.0.0"]
+build-backend = "poetry.masonry.api"
