### 2023-04-25 #5402 @EdyVision
>
Hello @gradvohl,
Thank you for your review. I am a little confused and concerned as the guidelines stated we can submit to JOSS using a feature branch with our paper (instead of having it in the main branch), and I couldn’t find the reference for the need to put the author/affiliations in the readme file. I do have my affiliations in the citations file and in the paper PR the review is attached to. If needed, I can merge the paper PR into the main branch, I just wasn’t prepared to do so in case there were updates requested to the paper. I have a PR to add the community guidelines from the readme into the `CONTRIBUTING.md` file and some other upgrades. I’ll have to check out Sphynx to expand on the documentation. Thank you for that suggestion. As for the install issue, I have a few other projects using this that cannot be open-sourced yet (they are due to be publicized this fall after my defense). I do have a notebook in [Google Collab](https://colab.research.google.com/gist/EdyVision/802ce21aab21eb5d9afa9e43d301eef7/pii-codex-sample-notebook.ipynb) where I was able to install and use the package. I’m linking the notebook to the readme for transparency, and I’ve stated in the readme that, at this time, only Ubuntu and MacOS were confirmed to work. Running locally, however, will require the use of poetry. I did update those instructions to be more clear.

Can you clarify where you were hoping for more citations? The README briefly touches on the primary citations used to derive categories and the enum types, but papers associated with the problem space were contained in the paper. The Milne types and the risk continuum levels and our ability to map to those types/levels were the primary conceptual targets, but I’m happy to revisit them. I wanted to make sure I understood expectations ahead of merging the next round of updates.

Thanks again for the review and your time!
