### 2023-04-25 e65224d
Er.add documentation (#30)
* Added contributing.md file
Updated presidio analyzer and spacy dependencies

* Removed extra community guidelines

* Updated readme and citation for PII mapping

* Updated test OS list
