+++ mypy.ini
@@ -16,5 +16,8 @@ ignore_missing_imports = True
 [mypy-presidio_anonymizer.*]
 ignore_missing_imports = True
 
+[mypy-spacy.*]
+ignore_missing_imports = True
+
 [mypy-pandas.*]
 ignore_missing_imports = True
