+++ joss/paper.bib
@@ -31,7 +31,7 @@
     month = {12},
     title = {{pii-codex: a Python library for PII detection, categorization, and severity assessment}},
     version = {0.4.3},
-    year = {2022}
+    year = {2023}
 }
 @misc{microsoft_presidio,
     title={Microsoft/Presidio: Context Aware, pluggable and customizable data protection and Anonymization SDK for text and images},
