### 2023-04-25 #5402 @arfon
> Thank you for your review. I am a little confused and concerned as the guidelines stated we can submit to JOSS using a feature branch with our paper (instead of having it in the main branch), and I couldn’t find the reference for the need to put the author/affiliations in the readme file. I do have my affiliations in the citations file and in the paper PR the review is attached to. If needed, I can merge the paper PR into the main branch, I just wasn’t prepared to do so in case there were updates requested for the paper.

Thanks for flagging this @EdyVision. @gradvohl – we don't actually require authors to merge the paper into their main/default branch. This is actually why we support compiling the paper from a non-default branch. @EdyVision – I'm assuming you want the review to be of the code on the `main` branch though right?
