### 2023-04-26 #5402 @EdyVision
> Hi @EdyVision,
>
> Regarding the paper in the `main` branch, I apologize for my mistake. In fact, it is not a JOSS requirement and therefore there is no problem with it being in the alternate branch.
>
> About the authors' names in the README.md file, I consider it important that your name is there. As a user, I think it is essential to know who is the author of that software and who is responsible for that code. I know that this information is in `CITATION.cff`, but I think it is critical to find this information quickly in `README.md` file.
>
> In my view, similar reasoning applies to the community contributions guidelines. I know they are in the `CONTRIBUTING.md` file, but I think it is important that they are clear in the `README.md` file itself. Or at least a section in the `README.md` file that indicates that more information is in the `CONTRIBUTING.md` file.
>
> Regarding the notebook in Google Collab, I saw that everything is implemented there. However, I am left with the question: which branch should I evaluate? The main or the joss-paper-submission? I haven't received any official information about this. If you or the Editor (@arfon) could be so kind as to tell me which branch I should analyze, it would help me a lot and save me time.
>
> Regarding citations, the ones you cite in the README.md file -- **Milne et al. (2016)** and **Schwartz and Solove (2012)** -- should be referenced in the `README.md` itself. After all, if a user sees these citations, eventually he will want to look them up for more information.
>
> Anyway, I don't know if I was able to answer your questions, but I remain at your disposal should you need further clarification.

Ah thank you @gradvohl, this helps! Ok so in my next pushes, I'll add the author details and point them to the `contributing.md` doc regarding the guidelines and such. All citations in the `README.md` file should exist in the `paper.md` file presently, but I'll make sure that it is explicitly noted if not. I'll review the paper this evening regarding the problem space and any associated citations in case something is missing or requires expansion. The branch being evaluated is the `joss-paper-submission` one and it is kept with the most up to date changes. I appreciate your time, thank you for the feedback and clarifications. I'll get started on these smaller items ahead of the documentation updates.
