### 2023-04-26 728a5d4
Added community guidelines blurb and author entry for readme
