### 2023-04-27 98b49c6
Added citations to readme and added missing foundational citations.
