+++ joss/paper.bib
@@ -111,3 +111,28 @@
     YEAR   = {2017},
     Note   = {To appear}
 }
+
+@article{belanger_crossler_2011,
+    ISSN = {02767783},
+    URL = {http://www.jstor.org/stable/41409971},
+    abstract = {Information privacy refers to the desire of individuals to control or have some influence over data about themselves. Advances in information technology have raised concerns about information privacy and its impacts, and have motivated Information Systems researchers to explore information privacy issues, including technical solutions to address these concerns. In this paper, we inform researchers about the current state of information privacy research in IS through a critical analysis of the IS literature that considers information privacy as a key construct. The review of the literature reveals that information privacy is a multilevel concept, but rarely studied as such. We also find that information privacy research has been heavily reliant on studentbased and USA-centric samples, which results in findings of limited generalizability. Information privacy research focuses on explaining and predicting theoretical contributions, with few studies in journal articles focusing on design and action contributions. We recommend that future research should consider different levels of analysis as well as multilevel effects of information privacy. We illustrate this with a multilevel framework for information privacy concerns. We call for research on information privacy to use a broader diversity of sampling populations, and for more design and action information privacy research to be published in journal articles that can result in IT artifacts for protection or control of information privacy.},
+    author = {France Bélanger and Robert E. Crossler},
+    journal = {MIS Quarterly},
+    number = {4},
+    pages = {1017--1041},
+    publisher = {Management Information Systems Research Center, University of Minnesota},
+    title = {Privacy in the Digital Age: A Review of Information Privacy Research in Information Systems},
+    urldate = {2023-04-26},
+    volume = {35},
+    year = {2011}
+}
+
+@article{tene_polonetsky_2012,
+    author = {Tene, Omer and Polonetsky, Jules},
+    year = {2012},
+    month = {09},
+    pages = {},
+    title = {Big Data for All: Privacy and User Control in the Age of Analytics},
+    volume = {11},
+    journal = {Northwestern Journal of Technology and Intellectual Property}
+}
