### 2023-04-27 7ece748
Added autogenerated docs and expanded doc strings for autogeneration
