+++ README.md
@@ -148,6 +148,7 @@ For more information on usage, check out the respective documentation for guidan
 | PII Detections and Analysis | [PII Detection and Analysis](docs/DETECTION_AND_ANALYSIS.md)          | Overview of how to detect and analyze strings                                            |
 | Local Repo Setup            | [Local Repo Setup](docs/LOCAL_SETUP.md)                               | Instructions for local repository setup                                                  |
 | Example Analysis            | [Example Analysis Notebook](notebooks/pii-analysis-ms-presidio.ipynb) | Notebook with example analysis using MSFT Presidio                                       |
+| PII-Codex Docs              | [PII-Codex Docs](docs/pii_codex/index.html)                           | Autogenerated docs on classes, services, and models                                      |
 
 <hr/>
 
