+++ CONTRIBUTING.md
@@ -21,6 +21,9 @@ make test.coverage
 
 Works in progress will be considered if you're unsure of what this exactly means, in which case you'll likely be asked to make some further changes.
 
+### Updating Documentation
+Docs are autogenerated by [pdoc](https://pdoc3.github.io/pdoc/). To update documentation, run the command `make docs` from terminal / command line. The generated docs will appear in docs/dev in html format. If the auto-generated doc entry is missing context, check the docstring added to the new code.
+
 ### Creating Releases
 With any change (non-breaking dependency upgrades, bug fixes, etc), the version will need to be updated to the next value and a new release created. Using the command listing below, run the one associated with the changes:
 
