+++ pyproject.toml
@@ -27,6 +27,7 @@ pillow = "^9.3.0"  # security dependency
 spacy = { version = "^3.5.2", optional = true }
 presidio-analyzer = { version = "^2.2.32", optional = true }
 presidio-anonymizer = { version ="^2.2.32", optional = true }
+pdoc = "^13.1.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.2.0"
