### 2023-05-29 #5402 @EdyVision
> Also, @EdyVision – do you have any updates on how you're getting on making updates based on @gradvohl's feedback?

Hello @arfon I've actually committed the changes requested by @gradvohl and updated the branch to contain the elements. I know that the largest item was that, for some reason, it couldn't be tested on the host described, but I cannot reproduce this, so I expanded the readme to include more details for setup. As for the mentions of the other projects, I defend them in a few weeks, so I believe I'll be able to expose those projects soon. I'll need to get confirmation from my advisor. That item aside, we are on track with the changes requested, and they are available in the branch. I'm happy to expand on the items if needed.

Thanks!
