### 2023-06-07 #5402 @tmickleydoyle
## Note about `A statement of need`

> The general knowledge base of identifiable data, the usage restrictions of this data, and the associated policies surrounding it have shifted drastically over the years.

I like the direction of this sentence, and I need to understand more about the shift you are seeing. Where were we ten years ago vs. now vs. five years in the future? In my opinion, saying `shifted drastically over the years` is too general for a statement of need,
