### 2023-06-16 #5402 @EdyVision
> ## Note about `A statement of need`
> _Not blocking_
>
> > The general knowledge base of identifiable data, the usage restrictions of this data, and the associated policies surrounding it have shifted drastically over the years.
>
> I like the direction of this sentence, and I need to understand more about the shift you are seeing. Where were we ten years ago vs. now vs. five years in the future? In my opinion, saying `shifted drastically over the years` is too general for a statement of need,

Hello @tmickleydoyle, thank you so much for the review. If I recall where I was going with this, its the general shift in policies and our approach to user awareness when it comes to privacy. Since before social media became a thing, netizens were openly blogging about experiences, places they've been, or sharing other details of themselves. It was a free for all. In a way, it still sort of is with social media. We openly share details of one another without considering the consequences. As more policies have started being enforced, the public is becoming more aware of their personal data disclosure activities and data rights.

Cloud service providers like Amazon Web Services have started providing a way to redact PII found in logs and such to help with possible exposures. Their PII identification API alone was introduced in March 2021 and I believe Microsoft Azure's capabilities was just a few years earlier.

I can restructure the sentence/paragraph, I see what you are saying. It needs more specificity on what the shift was, how it was historically and the sources for that information. I can work on that today and push the paper edits to the joss-paper-submission branch. I appreciate the feedback 🙏🏽
