### 2023-06-16 2377fb2
Updated statement of need in paper.md
