+++ joss/paper.bib
@@ -7,6 +7,24 @@
     author={Milne, George R. and Pettinico, George and Hajjat, Fatima M. and Markos, Ereni},
     year={2016}, pages={133–161}
 }
+@misc{OCR_2022,
+    title={Summary of the HIPAA privacy rule},
+    url={https://www.hhs.gov/hipaa/for-professionals/privacy/laws-regulations/index.html#:~:text=The%20U.S.%20Department%20of%20Health,1996%20(%22HIPAA%22).},
+    journal={HHS.gov},
+    author={(OCR), Office for Civil Rights},
+    year={2022},
+    month={Oct}
+}
+@article{West_2017,
+    title={Data capitalism: Redefining the logics of surveillance and privacy},
+    volume={58},
+    DOI={10.1177/0007650317718185},
+    number={1},
+    journal={Business &amp;amp; Society},
+    author={West, Sarah Myers},
+    year={2017},
+    pages={20–41}
+}
 @article{schwartz_solove_2011,
     title={The PII Problem: Privacy and a New Concept of Personally Identifiable Information},
     volume={86},
@@ -63,11 +81,17 @@
      author={Amazon Web Services},
      year={2022}
 }
-@misc{gdpr,
+@misc{gdpr_erasure_right,
     url={https://ico.org.uk/for-organisations/guide-to-data-protection/guide-to-the-general-data-protection-regulation-gdpr/individual-rights/right-to-erasure/#:~:text=Under%20Article%2017%20of%20the,be%20created%20in%20the%20future.},
     title={Right to erasure},
     publisher={Information Commissioners Office},
     year={2022}
+}
+@misc{GDPR_eu_2022,
+    url={https://gdpr.eu/what-is-gdpr/},
+    journal={GDPR.eu},
+    year={2022},
+    month={May}
 }
 @misc{hipaa,
     url={https://www.hhs.gov/hipaa/index.html},
