+++ joss/paper.bib
@@ -11,7 +11,7 @@
     title={Summary of the HIPAA privacy rule},
     url={https://www.hhs.gov/hipaa/for-professionals/privacy/laws-regulations/index.html#:~:text=The%20U.S.%20Department%20of%20Health,1996%20(%22HIPAA%22).},
     journal={HHS.gov},
-    author={(OCR), Office for Civil Rights},
+    author={OCR, Office for Civil Rights},
     year={2022},
     month={Oct}
 }
