### 2023-06-16 #5402 @EdyVision
@tmickleydoyle I've uploaded the revised Statement of Need to the branch with the new citations to help support it. I focused more on the policies driven by privacy-related events instead of the cloud service provider-specific implementations. Please let me know if these changes align with what you were seeking or if you would prefer the addition of CSP implementations to help guide this train of thought. Thanks again so much!
