### 2023-06-18 #5402 @gradvohl
Hello @arfon and @edyvision,

  I looked again in the repository, in the joss-paper-submission branch (actually, I never know precisely which branch I should look), and followed the instructions to run locally and install pii-codex. On the first instruction, it already shows an error. Maybe it is a case of mentioning https://spacy.io/usage and copying the spaCy installation information from there.

Then, even though I installed spaCy, the following errors occurred:
```console
pip install pii-codex
Defaulting to user installation because normal site-packages is not writeable
ERROR: Could not find a version that satisfies the requirement pii-codex (from versions: none)
ERROR: No matching distribution found for pii-codex
```

Therefore, I would ask @edyvision to check these steps or the requirements needed for the software to install and run.

Other aspects missed before, such as statement of need, references, community guidelines, contributions etc. are already fullfilled.
