### 2023-06-18 #5402 @EdyVision
> Hello @arfon and @EdyVision,
>
> I looked again in the repository, in the joss-paper-submission branch (actually, I never know precisely which branch I should look), and followed the instructions to run locally and install pii-codex. On the first instruction, it already shows an error. Maybe it is a case of mentioning https://spacy.io/usage and copying the spaCy installation information from there.
>
> Then, even though I installed spaCy, the following errors occurred:
>
> ```
> pip install pii-codex
> Defaulting to user installation because normal site-packages is not writeable
> ERROR: Could not find a version that satisfies the requirement pii-codex (from versions: none)
> ERROR: No matching distribution found for pii-codex
> ```
>
> Therefore, I would ask @EdyVision to check these steps or the requirements needed for the software to install and run.
>
> Other aspects missed before, such as statement of need, references, community guidelines, contributions etc. are already fullfilled.

Thanks @gradvohl. This is what I couldn't replicate. The branch for this submission was `joss-paper-submission`. The GH action steps install with Poetry, but my notebooks use pip, and I was able to install with just the following:

```
pip install --upgrade pip
pip install pii-codex
pip install pii-codex[detections]
```

spaCy and the Microsoft Presidio libs are in the extras and are installed as part of it. In the Google Collab notebook, I don't do any further installations outside of a `python -m spacy download en_core_web_lg` ahead of installing pii-codex.  I did note that only MacOS and Ubuntu have been used for testing.

from .github/workflows/test.yml file:
```
jobs:
  build:
    name: Run Tests
    strategy:
      matrix:
        python-version: [ "3.9", "3.10" ]
        os: [ubuntu-22.04, macos-latest]
    runs-on: ${{ matrix.os }}
    ```

I just added some other recommended entries from the spaCy site, so the readme entries for this section appear as follows:

```
pip install --upgrade pip
pip install -U pip setuptools wheel # only need if you haven't already done so
```

Before adding `pii-codex` on your project, download the spaCy `en_core_web_lg` model:

```
pip install -U spacy
python3 -m spacy download en_core_web_lg
```

These changes have been pushed up to the `joss-paper-submission` branch. By chance, do you encounter a similar error when using a hosted notebook? I'm trying to understand differences in setup or possible restrictions I'm not aware of.
