### 2023-06-18 #5402 @EdyVision
> Hello @EdyVision
> I don't use any notebooks. Could you provide a video with a demo of your software running?
> Perhaps it is easier for a user to follow the steps.

I can certainly do that. I'll work on that today and ping everyone when uploaded. Thanks!
