### 2023-06-18 #5402 @EdyVision
As requested, the README.md has been updated to note that within the LOCAL_SETUP.md, there is a video link of the walk-through (without sound - at this time, I cannot narrate videos). It shows new project creation from the virtual environment, installation of module, file creation, and finally, the execution of the file to show what the module is intended to do.

This video is an unlisted YouTube video (link share-able only): https://youtu.be/51TP2I5SNlo

@gradvohl @tmickleydoyle @arfon let me know if this helps.
