### 2023-06-18 #5402 @arfon
:wave: @EdyVision – some suggested edits to your paper:

- Please define PII the first time it's used. I would suggest ideally in the title and the first paragraph of the body of the paper. If you don't want to change the title, then please define at the start of the main paper body.
- The `West_2017` citation isn't rendering in the _Statement of Need_ section.
- Suggestion to change `utils` to `utilities` in the first sentence of the `Design` section.
