### 2023-06-18 #5402 @arfon
@EdyVision – Once you've made the suggested changes above, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
