### 2023-06-18 34cb417
Paper rendering edits
