### 2023-06-18 48f9e93
Joss paper submission (#28)
* JOSS Paper Submission

* Bib fix

* Added availability

* Added availability

* Updated paper with examples

* Fixed overflow issue with paper

* Added equations and content edits

* Citation fix

* Added pdf generator workflow

* Added JOSS badge to readme

* Removed HL3 license per JOSS requirement and added description in pyproject.toml

* Updated statement of need in paper.md

* Added community guidelines blurb and author entry for readme

* Added citations to readme and added missing foundational citations.

* Updated to checkoutv3

* Added autogenerated docs
[…]
