+++ joss/paper.bib
@@ -1,0 +1,162 @@
+@article{milne_pettinico_hajjat_markos_2016,
+    title={Information sensitivity typology: Mapping the degree and type of risk consumers perceive in personal data sharing},
+    volume={51},
+    DOI={10.1111/joca.12111},
+    number={1},
+    journal={Journal of Consumer Affairs},
+    author={Milne, George R. and Pettinico, George and Hajjat, Fatima M. and Markos, Ereni},
+    year={2016}, pages={133–161}
+}
+@misc{OCR_2022,
+    title={Summary of the HIPAA privacy rule},
+    url={https://www.hhs.gov/hipaa/for-professionals/privacy/laws-regulations/index.html#:~:text=The%20U.S.%20Department%20of%20Health,1996%20(%22HIPAA%22).},
+    journal={HHS.gov},
+    author={OCR, Office for Civil Rights},
+    year={2022},
+    month={Oct}
+}
+@article{West_2017,
+    title={Data capitalism: Redefining the logics of surveillance and privacy},
+    volume={58},
+    DOI={10.1177/0007650317718185},
+    number={1},
+    journal={Business &amp;amp; Society},
+    author={West, Sarah Myers},
+    year={2017},
+    pages={20–41}
+}
+@article{schwartz_solove_2011,
+    title={The PII Problem: Privacy and a New Concept of Personally Identifiable Information},
+    volume={86},
+    journal={New York University Law Review},
+    author={Schwartz, Paul M and Solove, Daniel J},
+    year={2011},
+    pages={1814}
+}
+@article{trepte2020,
+    author = {Trepte, Sabine},
+    year = {2020},
+    month = {05},
+    pages = {},
+    title = {The Social Media Privacy Model: Privacy and Communication in the Light of Social Media Affordances},
+    volume = {31},
+    journal = {Communication Theory},
+    doi = {10.1093/ct/qtz035}
+}
+@software{rosado2022,
+    author = {Rosado, Eidan J.},
+    doi = {10.5281/zenodo.7212576},
+    month = {12},
+    title = {{pii-codex: a Python library for PII detection, categorization, and severity assessment}},
+    version = {0.4.3},
+    year = {2023}
+}
+@misc{microsoft_presidio,
+    title={Microsoft/Presidio: Context Aware, pluggable and customizable data protection and Anonymization SDK for text and images},
+    url={https://github.com/microsoft/presidio},
+    publisher={Microsoft},
+    author={Microsoft}
+}
+@misc{microsoft_presidio_entities,
+    title={PII entities supported by Presidio},
+    url={https://microsoft.github.io/presidio/supported_entities/},
+    journal={Supported entities - Microsoft Presidio},
+    publisher={Microsoft},
+    author={Microsoft},
+    year={2022}
+}
+@misc{azure_detection_cognitive_skill,
+     title={PII detection cognitive skill - azure cognitive search},
+     url={https://learn.microsoft.com/en-us/azure/search/cognitive-search-skill-pii-detection},
+     journal={PII Detection cognitive skill - Azure Cognitive Search | Microsoft Learn},
+     publisher={Microsoft Azure},
+     author={Microsoft Azure},
+     year={2022}
+}
+@misc{aws_comprehend,
+     title={PII detection cognitive skill - azure cognitive search},
+     url={https://docs.aws.amazon.com/comprehend/latest/dg/what-is.html},
+     journal={What is Amazon Comprehend | AWS Comprehend},
+     publisher={Amazon Web Services},
+     author={Amazon Web Services},
+     year={2022}
+}
+@misc{gdpr_erasure_right,
[…]
