### 2023-06-18 #5402 @EdyVision
@arfon I've confirmed the Zenodo title matches the paper title. For this reason, the title itself was not updated, but the first sentence in the summary, I made sure that PII is accompanied by `personal identifiable information`. In the design section, utils has been changed to utilities as requested, and the citation for West 2017 has been corrected as of this last proof.

Version bumped to `0.4.5` for official paper release.

GitHub (with release notes): https://github.com/EdyVision/pii-codex/releases/tag/v0.4.5
Pypi: https://pypi.org/project/pii-codex/
Zenodo: https://zenodo.org/record/8053024

Zenodo DOI for v0.4.5: https://doi.org/10.5281/zenodo.8053024
Zenodo DOI for all versions: https://doi.org/10.5281/zenodo.7170516

Thanks so much, everyone!
