### 2023-06-18 d8f1a84
Er.fix release workflow (#32)
* Removed success check from workflow

* citation fixes
