+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
     given-names: Eidan J.
     orcid: https://orcid.org/0000-0003-0665-098X
 title: "pii-codex: a Python library for PII detection, categorization, and severity assessment"
-version: 0.4.4
+version: 0.4.6
 doi: 10.5281/zenodo.7212576
 date-released: 2023-06-18
